feat: Insert the table orders and projects

+++ LearningCenter.API/Security/Persistence/Repositories/UserRepository.cs
@@ -23,7 +23,7 @@ public class UserRepository : BaseRepository, IUserRepository
         await _context.Users.AddAsync(user);
     }
 
-    public async Task<User> FindByIdAsync(int id)
+    public async Task<User> FindByIdAsync(long id)
     {
         return await _context.Users.FindAsync(id);
     }
@@ -38,7 +38,7 @@ public class UserRepository : BaseRepository, IUserRepository
         return _context.Users.Any(x => x.Username == username);
     }
 
-    public User FindById(int id)
+    public User FindById(long id)
     {
         return _context.Users.Find(id);
     }
